+++ power_grid_model_c/power_grid_model/include/power_grid_model/main_model_impl.hpp
@@ -7,6 +7,7 @@
 // main model class
 
 // main include
+#include "batch_dispatch.hpp"
 #include "batch_parameter.hpp"
 #include "calculation_parameters.hpp"
 #include "container.hpp"
@@ -41,7 +42,6 @@
 // stl library
 #include <memory>
 #include <span>
-#include <thread>
 
 namespace power_grid_model {
 
@@ -147,18 +147,23 @@ class MainModelImpl<ExtraRetrievableTypes<ExtraRetrievableType...>, ComponentLis
     // container class
     using ComponentContainer = Container<ExtraRetrievableTypes<ExtraRetrievableType...>, ComponentType...>;
     using MainModelState = main_core::MainModelState<ComponentContainer>;
-    using MathState = main_core::MathState;
 
     using SequenceIdxView = std::array<std::span<Idx2D const>, main_core::utils::n_types<ComponentType...>>;
     using OwnedUpdateDataset = std::tuple<std::vector<typename ComponentType::UpdateType>...>;
 
-    static constexpr Idx ignore_output{-1};
+    using BatchDispatcher =
+        BatchDispatch<MainModelImpl<ExtraRetrievableTypes<ExtraRetrievableType...>, ComponentList<ComponentType...>>,
+                      ComponentType...>;
+
+    static constexpr Idx ignore_output{BatchDispatcher::sequential};
     static constexpr Idx isolated_component{-1};
     static constexpr Idx not_connected{-1};
-    static constexpr Idx sequential{-1};
+    static constexpr Idx sequential{BatchDispatcher::sequential};
 
   public:
     using Options = MainModelOptions;
+    using MathState = main_core::MathState;
+    using MetaData = meta_data::MetaData;
 
     // constructor with data
     explicit MainModelImpl(double system_frequency, ConstDataset const& input_data,
@@ -178,7 +183,6 @@ class MainModelImpl<ExtraRetrievableTypes<ExtraRetrievableType...>, ComponentLis
           meta_data_{&meta_data},
           math_solver_dispatcher_{&math_solver_dispatcher} {}
 
-  private:
     // helper function to get what components are present in the update data
     std::array<bool, main_core::utils::n_types<ComponentType...>>
     get_components_to_update(ConstDataset const& update_data) const {
@@ -188,6 +192,7 @@ class MainModelImpl<ExtraRetrievableTypes<ExtraRetrievableType...>, ComponentLis
             });
     }
 
+  private:
     // helper function to add vectors of components
     template <class CompType> void add_component(std::vector<typename CompType::InputType> const& components) {
         add_component<CompType>(components.begin(), components.end());
@@ -285,6 +290,7 @@ class MainModelImpl<ExtraRetrievableTypes<ExtraRetrievableType...>, ComponentLis
         }
     }
 
+  public:
     // overload to update all components across all scenarios
     template <cache_type_c CacheType, typename SequenceIdxMap>
         requires(std::same_as<SequenceIdxMap, main_core::utils::SequenceIdx<ComponentType...>> ||
@@ -298,7 +304,6 @@ class MainModelImpl<ExtraRetrievableTypes<ExtraRetrievableType...>, ComponentLis
             });
     }
 
-  public:
     // overload to update all components in the first scenario (e.g. permanent update)
     template <cache_type_c CacheType> void update_components(ConstDataset const& update_data) {
         auto const components_to_update = get_components_to_update(update_data);
@@ -394,6 +399,7 @@ class MainModelImpl<ExtraRetrievableTypes<ExtraRetrievableType...>, ComponentLis
         }
     }
 
+  public:
     // restore the initial values of all components
     void restore_components(SequenceIdxView const& sequence_idx) {
         (restore_component<ComponentType>(sequence_idx), ...);
@@ -401,6 +407,8 @@ class MainModelImpl<ExtraRetrievableTypes<ExtraRetrievableType...>, ComponentLis
         update_state(cached_state_changes_);
         cached_state_changes_ = {};
     }
+
+  private:
     void restore_components(std::array<std::reference_wrapper<std::vector<Idx2D> const>,
                                        main_core::utils::n_types<ComponentType...>> const& sequence_idx) {
         restore_components(std::array{std::span<Idx2D const>{
@@ -502,237 +510,8 @@ class MainModelImpl<ExtraRetrievableTypes<ExtraRetrievableType...>, ComponentLis
         requires std::invocable<std::remove_cvref_t<Calculate>, MainModelImpl&, MutableDataset const&, Idx>
     BatchParameter batch_calculation_(Calculate&& calculation_fn, MutableDataset const& result_data,
                                       ConstDataset const& update_data, Idx threading = sequential) {
-        return batch_calculation_(*this, calculation_info_, std::forward<Calculate>(calculation_fn), result_data,
-                                  update_data, threading);
-    }
-    template <typename Calculate>
-        requires std::invocable<std::remove_cvref_t<Calculate>, MainModelImpl&, MutableDataset const&, Idx>
-    static BatchParameter batch_calculation_(MainModelImpl& model, CalculationInfo& calculation_info,
-                                             Calculate&& calculation_fn, MutableDataset const& result_data,
-                                             ConstDataset const& update_data, Idx threading = sequential) {
-        // if the update dataset is empty without any component
-        // execute one power flow in the current instance, no batch calculation is needed
-        if (update_data.empty()) {
-            std::forward<Calculate>(calculation_fn)(model, result_data, 0);
-            return BatchParameter{};
-        }
-
-        // get batch size
-        Idx const n_scenarios = update_data.batch_size();
-
-        // if the batch_size is zero, it is a special case without doing any calculations at all
-        // we consider in this case the batch set is independent and but not topology cachable
-        if (n_scenarios == 0) {
-            return BatchParameter{};
-        }
-
-        // calculate once to cache topology, ignore results, all math solvers are initialized
-        try {
-            calculation_fn(model,
-                           {
-                               false,
-                               1,
-                               "sym_output",
-                               *model.meta_data_,
-                           },
-                           ignore_output);
-        } catch (SparseMatrixError const&) { // NOLINT(bugprone-empty-catch) // NOSONAR
-            // missing entries are provided in the update data
-        } catch (NotObservableError const&) { // NOLINT(bugprone-empty-catch) // NOSONAR
-            // missing entries are provided in the update data
-        }
-
-        // error messages
-        std::vector<std::string> exceptions(n_scenarios, "");
-        std::vector<CalculationInfo> infos(n_scenarios);
-
-        // lambda for sub batch calculation
-        main_core::utils::SequenceIdx<ComponentType...> all_scenarios_sequence;
-        auto sub_batch = sub_batch_calculation_(model, std::forward<Calculate>(calculation_fn), result_data,
-                                                update_data, all_scenarios_sequence, exceptions, infos);
-
-        batch_dispatch(sub_batch, n_scenarios, threading);
-
-        handle_batch_exceptions(exceptions);
-        calculation_info = main_core::merge_calculation_info(infos);
-
-        return BatchParameter{};
-    }
-
-    template <typename Calculate>
-        requires std::invocable<std::remove_cvref_t<Calculate>, MainModelImpl&, MutableDataset const&, Idx>
-    static auto sub_batch_calculation_(MainModelImpl const& base_model, Calculate&& calculation_fn,
-                                       MutableDataset const& result_data, ConstDataset const& update_data,
-                                       main_core::utils::SequenceIdx<ComponentType...>& all_scenarios_sequence,
-                                       std::vector<std::string>& exceptions, std::vector<CalculationInfo>& infos) {
-        // cache component update order where possible.
-        // the order for a cacheable (independent) component by definition is the same across all scenarios
-        auto const components_to_update = base_model.get_components_to_update(update_data);
-        auto const update_independence = main_core::update::independence::check_update_independence<ComponentType...>(
-            base_model.state_, update_data);
-        all_scenarios_sequence = main_core::update::get_all_sequence_idx_map<ComponentType...>(
-            base_model.state_, update_data, 0, components_to_update, update_independence, false);
-
-        return [&base_model, &exceptions, &infos, calculation_fn_ = std::forward<Calculate>(calculation_fn),
-                &result_data, &update_data, &all_scenarios_sequence_ = std::as_const(all_scenarios_sequence),
-                components_to_update, update_independence](Idx start, Idx stride, Idx n_scenarios) {
-            assert(n_scenarios <= narrow_cast<Idx>(exceptions.size()));
-            assert(n_scenarios <= narrow_cast<Idx>(infos.size()));
-
-            Timer const t_total(infos[start], 0000, "Total in thread");
-
-            auto const copy_model_functor = [&base_model, &infos](Idx scenario_idx) {
-                Timer const t_copy_model_functor(infos[scenario_idx], 1100, "Copy model");
-                return MainModelImpl{base_model};
-            };
-            auto model = copy_model_functor(start);
-
-            auto current_scenario_sequence_cache = main_core::utils::SequenceIdx<ComponentType...>{};
-            auto [setup, winddown] =
-                scenario_update_restore(model, update_data, components_to_update, update_independence,
-                                        all_scenarios_sequence_, current_scenario_sequence_cache, infos);
-
-            auto calculate_scenario = MainModelImpl::call_with<Idx>(
-                [&model, &calculation_fn_, &result_data, &infos](Idx scenario_idx) {
-                    calculation_fn_(model, result_data, scenario_idx);
-                    infos[scenario_idx].merge(model.calculation_info_);
-                },
-                std::move(setup), std::move(winddown), scenario_exception_handler(model, exceptions, infos),
-                [&model, &copy_model_functor](Idx scenario_idx) { model = copy_model_functor(scenario_idx); });
-
-            for (Idx scenario_idx = start; scenario_idx < n_scenarios; scenario_idx += stride) {
-                Timer const t_total_single(infos[scenario_idx], 0100, "Total single calculation in thread");
-
-                calculate_scenario(scenario_idx);
-            }
-        };
-    }
-
-    // run sequential if
-    //    specified threading < 0
-    //    use hardware threads, but it is either unknown (0) or only has one thread (1)
-    //    specified threading = 1
-    template <typename RunSubBatchFn>
-        requires std::invocable<std::remove_cvref_t<RunSubBatchFn>, Idx /*start*/, Idx /*stride*/, Idx /*n_scenarios*/>
-    static void batch_dispatch(RunSubBatchFn sub_batch, Idx n_scenarios, Idx threading) {
-        // run batches sequential or parallel
-        auto const hardware_thread = static_cast<Idx>(std::thread::hardware_concurrency());
-        if (threading < 0 || threading == 1 || (threading == 0 && hardware_thread < 2)) {
-            // run all in sequential
-            sub_batch(0, 1, n_scenarios);
-        } else {
-            // create parallel threads
-            Idx const n_thread = std::min(threading == 0 ? hardware_thread : threading, n_scenarios);
-            std::vector<std::thread> threads;
-            threads.reserve(n_thread);
-            for (Idx thread_number = 0; thread_number < n_thread; ++thread_number) {
-                // compute each sub batch with stride
-                threads.emplace_back(sub_batch, thread_number, n_thread, n_scenarios);
-            }
-            for (auto& thread : threads) {
-                thread.join();
-            }
-        }
-    }
-
-    template <typename... Args, typename RunFn, typename SetupFn, typename WinddownFn, typename HandleExceptionFn,
-              typename RecoverFromBadFn>
-        requires std::invocable<std::remove_cvref_t<RunFn>, Args const&...> &&
-                 std::invocable<std::remove_cvref_t<SetupFn>, Args const&...> &&
-                 std::invocable<std::remove_cvref_t<WinddownFn>, Args const&...> &&
-                 std::invocable<std::remove_cvref_t<HandleExceptionFn>, Args const&...> &&
-                 std::invocable<std::remove_cvref_t<RecoverFromBadFn>, Args const&...>
-    static auto call_with(RunFn run, SetupFn setup, WinddownFn winddown, HandleExceptionFn handle_exception,
-                          RecoverFromBadFn recover_from_bad) {
-        return [setup_ = std::move(setup), run_ = std::move(run), winddown_ = std::move(winddown),
-                handle_exception_ = std::move(handle_exception),
-                recover_from_bad_ = std::move(recover_from_bad)](Args const&... args) {
-            try {
-                setup_(args...);
-                run_(args...);
-                winddown_(args...);
-            } catch (...) {
-                handle_exception_(args...);
-                try {
-                    winddown_(args...);
-                } catch (...) {
-                    recover_from_bad_(args...);
-                }
-            }
-        };
-    }
-
-    static auto scenario_update_restore(
-        MainModelImpl& model, ConstDataset const& update_data,
-        main_core::utils::ComponentFlags<ComponentType...> const& components_to_store,
-        main_core::update::independence::UpdateIndependence<ComponentType...> const& do_update_cache,
-        main_core::utils::SequenceIdx<ComponentType...> const& all_scenario_sequence,
-        main_core::utils::SequenceIdx<ComponentType...>& current_scenario_sequence_cache,
-        std::vector<CalculationInfo>& infos) noexcept {
-        main_core::utils::ComponentFlags<ComponentType...> independence_flags{};
-        std::ranges::transform(do_update_cache, independence_flags.begin(),
-                               [](auto const& comp) { return comp.is_independent(); });
-        auto const scenario_sequence = [&all_scenario_sequence, &current_scenario_sequence_cache,
-                                        independence_flags_ = std::move(independence_flags)]() -> SequenceIdxView {
-            return main_core::utils::run_functor_with_all_types_return_array<ComponentType...>(
-                [&all_scenario_sequence, &current_scenario_sequence_cache, &independence_flags_]<typename CT>() {
-                    constexpr auto comp_idx = main_core::utils::index_of_component<CT, ComponentType...>;
-                    if (std::get<comp_idx>(independence_flags_)) {
-                        return std::span<Idx2D const>{std::get<comp_idx>(all_scenario_sequence)};
-                    }
-                    return std::span<Idx2D const>{std::get<comp_idx>(current_scenario_sequence_cache)};
-                });
-        };
-
-        return std::make_pair(
-            [&model, &update_data, scenario_sequence, &current_scenario_sequence_cache, &components_to_store,
-             do_update_cache_ = std::move(do_update_cache), &infos](Idx scenario_idx) {
-                Timer const t_update_model(infos[scenario_idx], 1200, "Update model");
-                current_scenario_sequence_cache = main_core::update::get_all_sequence_idx_map<ComponentType...>(
-                    model.state_, update_data, scenario_idx, components_to_store, do_update_cache_, true);
-
-                model.template update_components<cached_update_t>(update_data, scenario_idx, scenario_sequence());
-            },
-            [&model, scenario_sequence, &current_scenario_sequence_cache, &infos](Idx scenario_idx) {
-                Timer const t_update_model(infos[scenario_idx], 1201, "Restore model");
-
-                model.restore_components(scenario_sequence());
-                std::ranges::for_each(current_scenario_sequence_cache,
-                                      [](auto& comp_seq_idx) { comp_seq_idx.clear(); });
-            });
-    }
-
-    // Lippincott pattern
-    static auto scenario_exception_handler(MainModelImpl& model, std::vector<std::string>& messages,
-                                           std::vector<CalculationInfo>& infos) {
-        return [&model, &messages, &infos](Idx scenario_idx) {
-            std::exception_ptr const ex_ptr = std::current_exception();
-            try {
-                std::rethrow_exception(ex_ptr);
-            } catch (std::exception const& ex) {
-                messages[scenario_idx] = ex.what();
-            } catch (...) {
-                messages[scenario_idx] = "unknown exception";
-            }
-            infos[scenario_idx].merge(model.calculation_info_);
-        };
-    }
-
-    static void handle_batch_exceptions(std::vector<std::string> const& exceptions) {
-        std::string combined_error_message;
-        IdxVector failed_scenarios;
-        std::vector<std::string> err_msgs;
-        for (Idx batch = 0; batch < static_cast<Idx>(exceptions.size()); ++batch) {
-            // append exception if it is not empty
-            if (!exceptions[batch].empty()) {
-                combined_error_message += "Error in batch #" + std::to_string(batch) + ": " + exceptions[batch];
-                failed_scenarios.push_back(batch);
-                err_msgs.push_back(exceptions[batch]);
-            }
-        }
-        if (!combined_error_message.empty()) {
-            throw BatchCalculationError(combined_error_message, failed_scenarios, err_msgs);
-        }
+        return BatchDispatcher::batch_calculation_(*this, calculation_info_, std::forward<Calculate>(calculation_fn),
+                                                   result_data, update_data, threading);
     }
 
     // Calculate with optimization, e.g., automatic tap changer
@@ -805,6 +584,15 @@ class MainModelImpl<ExtraRetrievableTypes<ExtraRetrievableType...>, ComponentLis
     }
 
     CalculationInfo calculation_info() const { return calculation_info_; }
+    auto const& state() const {
+        assert(construction_complete_);
+        return state_;
+    }
+    auto const& meta_data() const {
+        assert(construction_complete_);
+        assert(meta_data_ != nullptr);
+        return *meta_data_;
+    }
 
     void check_no_experimental_features_used(Options const& options) const {
         if (options.calculation_type == CalculationType::state_estimation &&
@@ -852,7 +640,7 @@ class MainModelImpl<ExtraRetrievableTypes<ExtraRetrievableType...>, ComponentLis
                                                // may be changed in const functions for metrics
 
     double system_frequency_;
-    meta_data::MetaData const* meta_data_;
+    MetaData const* meta_data_;
     MathSolverDispatcher const* math_solver_dispatcher_;
 
     MainModelState state_;
